Feat/add form toggle button (#321)
* Chore: install dependencies

* Feat: add FormFieldToggle

* Feat: change use of FormFieldDropdown in settings to FormFieldToggle instead

* Chore: delete unused FormFieldDropdown class

* Fix: typo

* Fix: add extra padding for toggle buttons

* Fix: shift toggle button for reach to bottom of footer settings
