+++ package.json
@@ -32,6 +32,7 @@
     "react-select": "^3.1.0",
     "react-simplemde-editor": "^4.1.0",
     "react-toastify": "^6.1.0",
+    "react-toggle-button": "^2.2.0",
     "sgds-govtech": "^1.3.16",
     "slugify": "^1.3.6",
     "toml": "^3.0.0",
