+++ src/styles/isomer-cms/elements/variable.scss
@@ -6,6 +6,7 @@ $dark-greyish-blue: #1B2E56;
 $body-text: #333;
 $body-grey: #878787;
 $login-green: #259a36;
+$success: #00875F;
 
 $default-background: white;
 $base-background: #E5E9F1;
